Add data persistence to the example project (fixes).

+++ api/src/main/java/de/unistuttgart/iste/ese/api/cats/CatRepository.java
@@ -2,7 +2,7 @@ package de.unistuttgart.iste.ese.api.cats;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CatRepository extends CrudRepository<Cat, Integer> {
+public interface CatRepository extends CrudRepository<Cat, Long> {
     Cat findByName(String name);
 
     Cat findById(long id);
